Extract JoinRoomComponent
