+++ client/src/index.ts
@@ -2,5 +2,5 @@ import { RootComponent } from "./Components/RootComponent";
 import { View } from "./View";
 
 document.addEventListener('DOMContentLoaded', function() {
-    (new View("root")).setView(new RootComponent());
+    (new View("app")).setView(new RootComponent());
 }, false);
